% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHTS2009.R
\docType{data}
\name{Hh_df}
\alias{Hh_df}
\title{Household data frame from the 2009 NTHS data.}
\format{The main data frame \code{Hh_df} has 130509 rows and 78 variables:
\describe{
  \item{AADVMT}{}
  \item{AADVMT.int}{}
  \item{Age0to14}{}
  \item{Age65Plus}{}
  \item{ANNMILES}{}
  \item{atd.miles.Auto}{}
  \item{atd.miles.Bike}{}
  \item{atd.miles.Transit}{}
  \item{atd.miles.Walk}{}
  \item{BESTMILE}{}
  \item{BESTMILEcap}{}
  \item{CENSUS_D}{}
  \item{CENSUS_R}{}
  \item{Drivers}{}
  \item{DrvAgePop}{}
  \item{DVMT}{}
  \item{DVMTcap}{}
  \item{DVMT.int}{}
  \item{FLAG100}{}
  \item{HBHTNRNT}{}
  \item{HBHUR}{}
  \item{HHFAMINC}{}
  \item{HHFAMINC10k}{}
  \item{HHFAMINC20k}{}
  \item{HHFAMINCVAL}{}
  \item{HhSize}{}
  \item{HOUSEID}{}
  \item{HTHTNRNT}{}
  \item{HTPPOPDN}{}
  \item{LifeCycle}{}
  \item{lnBESTMILE}{}
  \item{lntd.milescap}{}
  \item{LogIncome}{}
  \item{MSACAT}{}
  \item{MSASIZE}{}
  \item{ntrips}{}
  \item{ntrips.Auto}{}
  \item{ntrips.Bike}{}
  \item{ntrips.Transit}{}
  \item{ntrips.Walk}{}
  \item{NUMADLT}{}
  \item{powBESTMILE}{}
  \item{powDVMT}{}
  \item{RAIL}{}
  \item{TDAYDATE}{}
  \item{td.miles}{}
  \item{td.miles.Auto}{}
  \item{td.miles.Bike}{}
  \item{td.milescap}{}
  \item{td.miles.Transit}{}
  \item{td.miles.Walk}{}
  \item{td.pct.Auto}{}
  \item{td.pct.Bike}{}
  \item{td.pct.Transit}{}
  \item{td.pct.Walk}{}
  \item{TRAVDAY}{}
  \item{tt.mins}{}
  \item{tt.mins.Auto}{}
  \item{tt.mins.Bike}{}
  \item{tt.mins.Transit}{}
  \item{tt.mins.Walk}{}
  \item{tt.pct.Auto}{}
  \item{tt.pct.Bike}{}
  \item{tt.pct.Transit}{}
  \item{tt.pct.Walk}{}
  \item{URBAN}{}
  \item{URBANSIZE}{}
  \item{URBRUR}{}
  \item{Vehicles}{}
  \item{VehPerDriver}{}
  \item{VehPerDrvAgePop}{}
  \item{WEEKDAY}{}
  \item{WEEKEND}{}
  \item{Workers}{}
  \item{WTHHFIN}{}
  \item{ZeroAADVMT}{}
  \item{ZeroDVMT}{}
  \item{ZeroVeh}{}
  }}
\source{
\url{http://nhts.ornl.gov/download.shtml#2009}
}
\usage{
Hh_df
}
\description{
Household data frame from the 2009 NTHS data..
}
\examples{
str(Hh_df)
head(Hh_df)
summary(Hh_df)

}
\keyword{datasets}
